Stop DDOSing the public ecr to get latest versions
Made to prevent: `GET
https://public.ecr.aws/v2/dynatrace/dynatrace-codemodules/tags/list?n=1000:
TOOMANYREQUESTS: Rate exceeded`
Every test that needed the latest image version of a compononent pinged
the registry to do so. This ping happened no matter if the test was
supposed to run or not. This is because of the way the e2e framework
works(bunch of callbacks), so these requests happened all at once.