Fix implementation and revert unit test changes
My implementation was wrong which led me to a wrong path. Now I reverted the statefulset unit test with the additional manageStatefulSet call. BUT I removed one test that I added, which was about updating a configmap query. I think my test wast wrong and did not really update the configMap properly, as annotations are not immutable objects in K8S, in other words it does not lead to updating the object properly.